Added random number generator
the drawString is used to simple just test to see if it works (it does). just need to make a system where the timer (like what you did to the ball ) is used to set a time when a powerup should spawn.  It then will draw a box that used the xGenerator and yGenerator as the X and Y value. then  you can guess the rest.